+++ packages/core/src/authFlows/__tests__/adfs.unit.spec.ts
@@ -1,7 +1,7 @@
 import nock from 'nock';
 import noop from 'lodash/noop';
 import { ADFS } from '../adfs';
-import * as loginUtils from '../loginUtils';
+import * as loginUtils from '../../loginUtils';
 
 describe('ADFS', () => {
   const cluster = 'test-cluster';
