+++ packages/core/src/index.ts
@@ -38,7 +38,7 @@ export {
   AuthenticateParams,
   IdInfo,
   OnAuthenticateLoginObject,
-} from './auth';
+} from './authFlows/legacy';
 export {
   AuthorizeOptions,
   AuthorizeParams,
@@ -61,6 +61,6 @@ export {
   AzureADSignInType,
   AzureADSingInFlow,
   AzureADSignInRequestParams,
-} from './aad';
-export { ADFSRequestParams } from './adfs';
+} from './authFlows/aad';
+export { ADFSRequestParams } from './authFlows/adfs';
 export { TestUtils, Constants, GraphUtils };
